+++ apps/edge/recoverage-server/package.json
@@ -11,8 +11,10 @@
 		"db:up": "na wrangler d1 migrations apply recoverage-beta"
 	},
 	"dependencies": {
+		"arktype": "2.1.2",
 		"drizzle-orm": "0.40.0",
 		"hono": "4.7.2",
+		"htmx.org": "2.0.4",
 		"nanoid": "5.1.2",
 		"octokit": "4.1.2"
 	},
@@ -21,12 +23,10 @@
 		"@cloudflare/workers-types": "4.20250224.0",
 		"@hono/vite-build": "1.3.0",
 		"@hono/vite-dev-server": "0.18.2",
-		"arktype": "2.1.2",
 		"bun-types": "1.2.4",
 		"cloudflare": "4.1.0",
 		"drizzle-kit": "0.30.5",
 		"eslint": "9.21.0",
-		"htmx.org": "2.0.4",
 		"vite": "6.2.0",
 		"vitest": "3.0.7",
 		"wrangler": "3.111.0"
