--- apps/edge/recoverage-server/src/soft-fetch.ts
@@ -1,3 +1,0 @@
-// export function gentleFetcher(cacheTime: number) : Promise<typeof fetch> {
-
-// }
