--- apps/edge/recoverage-server/static/style.css
@@ -1,1 +1,0 @@
-h1 { font-family: Arial, Helvetica, sans-serif; }
