+++ apps/edge/recoverage-server/vite.config.ts
@@ -8,7 +8,7 @@ export default defineConfig({
 		build(),
 		devServer({
 			adapter,
-			entry: `src/index.tsx`,
+			entry: `src/app.tsx`,
 		}),
 	],
 })
